tests/wpt/tests/content-security-policy/script-src-attr-elem/script-src-attr-blocked-src-allowed.html

<!DOCTYPE html>
<html>

<head>
    <meta http-equiv="Content-Security-Policy" content="script-src-attr 'none';
    script-src 'unsafe-inline' 'self';">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
</head>

<body>
    <script>
      var t = async_test("Should fire a security policy violation event");
      window.addEventListener('securitypolicyviolation', t.step_func_done(function(e) {
        assert_equals(e.violatedDirective, 'script-src-attr');
        assert_equals(e.blockedURI, 'inline');
      }));
    </script>

    <img src="../support/pass.png" onload="t.unreached_func('Should not have executed the inline handler')">
</body>

</html>
