LayoutTests/imported/w3c/web-platform-tests/content-security-policy/script-src-attr-elem/script-src-attr-blocked-src-allowed-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'script-src-attr'.



Harness Error (TIMEOUT), message = null

NOTRUN Should fire a security policy violation event
